LayoutTests/platform/mac/accessibility/aria-hidden-false-works-in-subtrees-expected.txt

This tests that a sub-tree within a node marked with aria-hidden=false will be exposed in the AX tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


AXRole: AXGroup
AXDescription: group0

  AXRole: AXGroup
  AXDescription: group2

    AXRole: AXButton
    AXDescription: button1

    AXRole: AXStaticText
    AXDescription:

PASS object.role is 'AXRole: AXStaticText'
PASS stringValue is 'AXValue: text3'
PASS object.role is 'AXRole: AXStaticText'
PASS stringValue is 'AXValue:     text4  '
PASS !object || !object.isValid is true
PASS object.role is 'AXRole: AXGroup'
PASS object.childrenCount is 0
PASS object.role is 'AXRole: AXGroup'
PASS object.childrenCount is 1
PASS object.childAtIndex(0).role is 'AXRole: AXButton'
Non-rendered iframe content should not be visible when aria-hidden=true. The first child should be a group and NOT static text.
iFrame child role: AXRole: AXGroup
PASS successfullyParsed is true

TEST COMPLETE
